+++ src/Uno.UI.Composition/AssemblyInfo.cs
@@ -16,6 +16,8 @@ using System.Runtime.InteropServices;
 [assembly: InternalsVisibleTo("SamplesApp.Wasm")]
 [assembly: InternalsVisibleTo("SamplesApp.Skia")]
 
+[assembly: InternalsVisibleTo("Uno.WinUI.Graphics2DSK")]
+
 [assembly: InternalsVisibleTo("UnoIslandsSamplesApp")]
 [assembly: InternalsVisibleTo("UnoIslandsSamplesApp.Skia")]
 [assembly: System.Reflection.AssemblyMetadata("IsTrimmable", "True")]
