+++ src/Uno.UI.sln
@@ -317,6 +317,8 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Uno.UI.Runtime.Skia.X11", "
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Uno.WinUI.Graphics3DGL", "AddIns\Uno.WinUI.Graphics3DGL\Uno.WinUI.Graphics3DGL.csproj", "{0F62DA75-6AD9-4F58-B69C-D63CA9053E34}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Uno.WinUI.Graphics2DSK", "AddIns\Uno.WinUI.Graphics2DSK\Uno.WinUI.Graphics2DSK.csproj", "{B1A17B3C-8B9D-40C8-8DBF-37CAAC7EB55E}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -5772,6 +5774,48 @@ Global
 		{0F62DA75-6AD9-4F58-B69C-D63CA9053E34}.Release|x64.Build.0 = Release|Any CPU
 		{0F62DA75-6AD9-4F58-B69C-D63CA9053E34}.Release|x86.ActiveCfg = Release|Any CPU
 		{0F62DA75-6AD9-4F58-B69C-D63CA9053E34}.Release|x86.Build.0 = Release|Any CPU
+		{B1A17B3C-8B9D-40C8-8DBF-37CAAC7EB55E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{B1A17B3C-8B9D-40C8-8DBF-37CAAC7EB55E}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{B1A17B3C-8B9D-40C8-8DBF-37CAAC7EB55E}.Debug|ARM.ActiveCfg = Debug|Any CPU
+		{B1A17B3C-8B9D-40C8-8DBF-37CAAC7EB55E}.Debug|ARM.Build.0 = Debug|Any CPU
+		{B1A17B3C-8B9D-40C8-8DBF-37CAAC7EB55E}.Debug|ARM64.ActiveCfg = Debug|Any CPU
+		{B1A17B3C-8B9D-40C8-8DBF-37CAAC7EB55E}.Debug|ARM64.Build.0 = Debug|Any CPU
+		{B1A17B3C-8B9D-40C8-8DBF-37CAAC7EB55E}.Debug|iPhone.ActiveCfg = Debug|Any CPU
+		{B1A17B3C-8B9D-40C8-8DBF-37CAAC7EB55E}.Debug|iPhone.Build.0 = Debug|Any CPU
+		{B1A17B3C-8B9D-40C8-8DBF-37CAAC7EB55E}.Debug|iPhoneSimulator.ActiveCfg = Debug|Any CPU
+		{B1A17B3C-8B9D-40C8-8DBF-37CAAC7EB55E}.Debug|iPhoneSimulator.Build.0 = Debug|Any CPU
+		{B1A17B3C-8B9D-40C8-8DBF-37CAAC7EB55E}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{B1A17B3C-8B9D-40C8-8DBF-37CAAC7EB55E}.Debug|x64.Build.0 = Debug|Any CPU
+		{B1A17B3C-8B9D-40C8-8DBF-37CAAC7EB55E}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{B1A17B3C-8B9D-40C8-8DBF-37CAAC7EB55E}.Debug|x86.Build.0 = Debug|Any CPU
+		{B1A17B3C-8B9D-40C8-8DBF-37CAAC7EB55E}.Release_NoSamples|Any CPU.ActiveCfg = Debug|Any CPU
+		{B1A17B3C-8B9D-40C8-8DBF-37CAAC7EB55E}.Release_NoSamples|Any CPU.Build.0 = Debug|Any CPU
+		{B1A17B3C-8B9D-40C8-8DBF-37CAAC7EB55E}.Release_NoSamples|ARM.ActiveCfg = Debug|Any CPU
+		{B1A17B3C-8B9D-40C8-8DBF-37CAAC7EB55E}.Release_NoSamples|ARM.Build.0 = Debug|Any CPU
+		{B1A17B3C-8B9D-40C8-8DBF-37CAAC7EB55E}.Release_NoSamples|ARM64.ActiveCfg = Debug|Any CPU
+		{B1A17B3C-8B9D-40C8-8DBF-37CAAC7EB55E}.Release_NoSamples|ARM64.Build.0 = Debug|Any CPU
+		{B1A17B3C-8B9D-40C8-8DBF-37CAAC7EB55E}.Release_NoSamples|iPhone.ActiveCfg = Debug|Any CPU
+		{B1A17B3C-8B9D-40C8-8DBF-37CAAC7EB55E}.Release_NoSamples|iPhone.Build.0 = Debug|Any CPU
+		{B1A17B3C-8B9D-40C8-8DBF-37CAAC7EB55E}.Release_NoSamples|iPhoneSimulator.ActiveCfg = Debug|Any CPU
+		{B1A17B3C-8B9D-40C8-8DBF-37CAAC7EB55E}.Release_NoSamples|iPhoneSimulator.Build.0 = Debug|Any CPU
+		{B1A17B3C-8B9D-40C8-8DBF-37CAAC7EB55E}.Release_NoSamples|x64.ActiveCfg = Debug|Any CPU
+		{B1A17B3C-8B9D-40C8-8DBF-37CAAC7EB55E}.Release_NoSamples|x64.Build.0 = Debug|Any CPU
+		{B1A17B3C-8B9D-40C8-8DBF-37CAAC7EB55E}.Release_NoSamples|x86.ActiveCfg = Debug|Any CPU
+		{B1A17B3C-8B9D-40C8-8DBF-37CAAC7EB55E}.Release_NoSamples|x86.Build.0 = Debug|Any CPU
+		{B1A17B3C-8B9D-40C8-8DBF-37CAAC7EB55E}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{B1A17B3C-8B9D-40C8-8DBF-37CAAC7EB55E}.Release|Any CPU.Build.0 = Release|Any CPU
+		{B1A17B3C-8B9D-40C8-8DBF-37CAAC7EB55E}.Release|ARM.ActiveCfg = Release|Any CPU
+		{B1A17B3C-8B9D-40C8-8DBF-37CAAC7EB55E}.Release|ARM.Build.0 = Release|Any CPU
+		{B1A17B3C-8B9D-40C8-8DBF-37CAAC7EB55E}.Release|ARM64.ActiveCfg = Release|Any CPU
+		{B1A17B3C-8B9D-40C8-8DBF-37CAAC7EB55E}.Release|ARM64.Build.0 = Release|Any CPU
+		{B1A17B3C-8B9D-40C8-8DBF-37CAAC7EB55E}.Release|iPhone.ActiveCfg = Release|Any CPU
+		{B1A17B3C-8B9D-40C8-8DBF-37CAAC7EB55E}.Release|iPhone.Build.0 = Release|Any CPU
+		{B1A17B3C-8B9D-40C8-8DBF-37CAAC7EB55E}.Release|iPhoneSimulator.ActiveCfg = Release|Any CPU
+		{B1A17B3C-8B9D-40C8-8DBF-37CAAC7EB55E}.Release|iPhoneSimulator.Build.0 = Release|Any CPU
+		{B1A17B3C-8B9D-40C8-8DBF-37CAAC7EB55E}.Release|x64.ActiveCfg = Release|Any CPU
+		{B1A17B3C-8B9D-40C8-8DBF-37CAAC7EB55E}.Release|x64.Build.0 = Release|Any CPU
+		{B1A17B3C-8B9D-40C8-8DBF-37CAAC7EB55E}.Release|x86.ActiveCfg = Release|Any CPU
+		{B1A17B3C-8B9D-40C8-8DBF-37CAAC7EB55E}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -5916,6 +5960,7 @@ Global
 		{93E6D033-52E2-4C2B-A715-A9FC1F609E3D} = {995C1054-AB61-42EE-9A17-C32155BD6D13}
 		{37441DE3-088B-4B63-A1E2-E70E39BF4222} = {416684CF-A4E3-4079-B380-3FF0B00E433C}
 		{0F62DA75-6AD9-4F58-B69C-D63CA9053E34} = {E872CD33-A455-4DC4-8A87-5FEEC1C39A44}
+		{B1A17B3C-8B9D-40C8-8DBF-37CAAC7EB55E} = {E872CD33-A455-4DC4-8A87-5FEEC1C39A44}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {9B2608F4-D82B-4B72-B399-33E822DF01D0}
